MonitorPlus: Rebrand as RolloutInfoWrapper
MonitorPlus added raw observations and rewards to the
`info['episode']` dict generated by its parent class Monitor.

PPO2 and other Stable Baselines algorithms hold on to the 100 most
recent values of `info['episode']`.

For environment with long episodes and large observations, using
MonitorPlus and PPO2 used an unnecessarily large amount of memory
(up to 1.50 GB per episode). This patch reduced memory usage by
converting MonitorPlus into a separate wrapper RolloutInfoWrapper
which uses `info['rollout']` instead.
